requirements.txt
telethon
blinker
certifi
cffi
charset-normalizer
click
dominate
Flask
Flask-Bootstrap
Flask-WTF
idna
itsdangerous
Jinja2
MarkupSafe
MouseInfo
numpy
PyAudio
pycparser
PyGetWindow
pyperclip
python-dotenv
pytweening
requests
rubicon-objc
SoundCard
SpeechRecognition
typing_extensions
urllib3
visitor
Werkzeug
WTForms
scrapegraphai
langchain_community
scrapegraphai
opencv-python
pyvirtualcam
flask-mail
flask-login
qrcode
pyotp
email_validator
sentence_transformers
PyPDF2
pywhatkit
python-telegram-bot
aiogram
flask_socketio
pytest
selenium